[T] Move AutoCheckTools.h to Tests/.
It is used by more than few tests now.
