+++ Tests/AutoCheckTools.h
